+++ common/src/main/java/com/zigythebird/playeranim/animation/keyframe/KeyframeLocation.java
@@ -6,10 +6,10 @@
 package com.zigythebird.playeranim.animation.keyframe;
 
 /**
- * A named pair object that stores a {@link Keyframe} and a double representing a temporally placed {@code Keyframe}
+ * A named pair object that stores a {@link Keyframe} and a float representing a temporally placed {@code Keyframe}
  *
  * @param keyframe The {@code Keyframe} at the tick time
  * @param startTick The animation tick time at the start of this {@code Keyframe}
  * @param tick The animation tick of the keyframe
  */
-public record KeyframeLocation<T extends Keyframe>(T keyframe, double startTick, float tick) { }
+public record KeyframeLocation<T extends Keyframe>(T keyframe, float startTick, float tick) { }
